+++ src/app/kibaApp.tsx
@@ -5,19 +5,13 @@ import { getIsRunningOnBrowser, IMultiAnyChildProps, useInitialization } from '@
 import styled from 'styled-components';
 
 import { ITheme, ThemeProvider } from '../theming';
+import { BackgroundView, IBackgroundConfig } from '../wrappers';
 import { GlobalCss } from './globalCss';
+import { Head, HeadRootProvider, IHeadRootProviderProps } from './headContext';
 import { resetCss } from './resetCss';
 
 import 'lazysizes';
 import 'lazysizes/plugins/attrchange/ls.attrchange';
-
-interface IKibaAppProps extends IMultiAnyChildProps {
-  theme: ITheme;
-  isRehydrating?: boolean;
-  isFullPageApp?: boolean;
-  extraGlobalCss?: string;
-  extraCss?: string;
-}
 
 interface IStyledMainViewProps {
   extraCss?: string;
@@ -25,15 +19,24 @@ interface IStyledMainViewProps {
 
 const StyledMainView = styled.div`
   min-height: 100vh;
-  /* NOTE(krishan711): the min-height doesn't propagate to children that have height:100% unless this is here (https://stackoverflow.com/questions/8468066) */
-  height: 1px;
 
   &.fullPage {
+    /* NOTE(krishan711): the min-height doesn't propagate to children that have height:100% unless this is here (https://stackoverflow.com/questions/8468066) */
+    height: 1px;
     min-height: 100%;
   }
 
   ${(props: IStyledMainViewProps): string => props.extraCss || ''};
 `;
+
+export interface IKibaAppProps extends IMultiAnyChildProps, IHeadRootProviderProps {
+  theme: ITheme;
+  isRehydrating?: boolean;
+  isFullPageApp?: boolean;
+  extraGlobalCss?: string;
+  extraCss?: string;
+  background?: IBackgroundConfig;
+}
 
 export const KibaApp = (props: IKibaAppProps): React.ReactElement => {
   // NOTE(krishan711): the default is false because if this is rehydrating it would be false on the server and needs to match.
@@ -51,12 +54,28 @@ export const KibaApp = (props: IKibaAppProps): React.ReactElement => {
         extraCss={props.extraGlobalCss}
         isFullPageApp={props.isFullPageApp}
       />
-      <StyledMainView
-        className={getClassName(isRunningOnBrowser ? 'js' : 'no-js', props.isFullPageApp && 'fullPage')}
-        extraCss={props.extraCss}
-      >
-        {props.children}
-      </StyledMainView>
+      <HeadRootProvider setHead={props.setHead}>
+        <Head headId='kiba-app'>
+          <link rel='preconnect' href='https://assets.evrpg.com' crossOrigin='anonymous' />
+          { Object.keys(props.theme.fonts || {}).map((fontKey: string, index: number): React.ReactElement => (
+            <React.Fragment key={index}>
+              <link href={props.theme.fonts[fontKey].url} rel='preload' as='style' />
+              {/* TODO(krishan711): the lazy loading doesn't work here */}
+              {/* <link href={theme.fonts[fontKey].url} rel='stylesheet' media='print' onLoad={((event: React.SyntheticEvent<HTMLLinkElement>): void => {(event.target as HTMLLinkElement).media = 'all'})} />
+              <noscript><link href={theme.fonts[fontKey].url} rel='stylesheet' /></noscript> */}
+              <link href={props.theme.fonts[fontKey].url} rel='stylesheet' />
+            </React.Fragment>
+          ))}
+        </Head>
+        <BackgroundView { ...props.background }>
+          <StyledMainView
+            className={getClassName(isRunningOnBrowser ? 'js' : 'no-js', props.isFullPageApp && 'fullPage')}
+            extraCss={props.extraCss}
+          >
+            {props.children}
+          </StyledMainView>
+        </BackgroundView>
+      </HeadRootProvider>
     </ThemeProvider>
   );
 };