+++ src/atoms/link/component.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 
 import { getClassName } from '@kibalabs/core';
+import { Link as CoreLink, useIsCoreRoutingEnabled } from '@kibalabs/core-react';
 import styled from 'styled-components';
 
 import { defaultComponentProps, IComponentProps, themeToCss, useBuiltTheme } from '../..';
@@ -45,17 +46,20 @@ export interface ILinkProps extends IComponentProps<ILinkTheme> {
 }
 
 export const Link = (props: ILinkProps): React.ReactElement => {
+  const isUsingCoreRouting = useIsCoreRoutingEnabled();
   const theme = useBuiltTheme('links', props.variant, props.theme);
   const shouldOpenSameTab = props.shouldOpenSameTab || (props.shouldOpenSameTab === undefined && props.target && (props.target.startsWith('#') || props.target.startsWith('/')));
   return (
+  // @ts-ignore: as prop doesn't match type required
     <StyledLink
       id={props.id}
       className={getClassName(Link.displayName, props.className, !props.isEnabled && 'disabled')}
       theme={theme}
       href={props.isEnabled ? props.target : undefined}
+      rel={'noopener'}
       tabIndex={props.tabIndex || 0}
       target={shouldOpenSameTab ? '_self' : '_blank'}
-      rel={'noopener'}
+      as={isUsingCoreRouting && shouldOpenSameTab ? CoreLink : undefined}
     >
       {props.text}
     </StyledLink>